Rename lib/solid/ to lib/models/
