+++ lib/identity.js
@@ -9,7 +9,7 @@ module.exports.loadExtendedProfile = loadExtendedProfile
 
 var graphUtil = require('./util/graph-util')
 var webClient = require('./web')
-var SolidProfile = require('./solid/profile')
+var SolidProfile = require('./models/profile')
 var vocab = require('./vocab')
 
 /**
