+++ lib/models/index-registration.js
