+++ lib/models/resource.js
