+++ lib/type-registry.js
@@ -13,7 +13,7 @@ module.exports.registerType = registerType
 module.exports.typeRegistryForClass = typeRegistryForClass
 module.exports.unregisterType = unregisterType
 
-var IndexRegistration = require('./solid/index-registration')
+var IndexRegistration = require('./models/index-registration')
 var rdf = require('./util/rdf-parser').rdflib
 var webClient = require('./web')
 var util = require('./util/web-util.js')
