+++ lib/web.js
@@ -5,8 +5,8 @@
  */
 var config = require('../config')
 var graphUtil = require('./util/graph-util')
-var SolidResponse = require('./solid/response')
-var SolidContainer = require('./solid/container')
+var SolidResponse = require('./models/response')
+var SolidContainer = require('./models/container')
 var vocab = require('./vocab')
 var XMLHttpRequest = require('./util/xhr')
 var HttpError = require('standard-http-error')
